Split code into three separate Python packages
