+++ .gitignore
@@ -1,8 +1,9 @@
 /.eggs/
-/eggs/
+/cluster_mgr_eggs/
+/config_mgr_eggs/
 /_env/
 *.pyc
 /.coverage
 /build/
 /htmlcov/
-/src/clearwater_cluster_manager.egg-info/
+/src/*.egg-info/
