+++ Makefile
@@ -13,8 +13,8 @@ X86_64_ONLY=0
 .DEFAULT_GOAL = deb
 
 .PHONY: test
-test: setup.py env
-	${PYTHON_BIN} setup.py test -v
+test: cluster_mgr_setup.py env
+	${PYTHON_BIN} cluster_mgr_setup.py test -v
 
 ${ENV_DIR}/bin/flake8: env
 	${ENV_DIR}/bin/pip install flake8
@@ -32,15 +32,15 @@ explain-style: ${ENV_DIR}/bin/flake8
 	${ENV_DIR}/bin/flake8 --select=E,W,C,N --show-pep8 --first --max-line-length=100 src/
 
 .PHONY: coverage
-coverage: ${ENV_DIR}/bin/coverage setup.py
+coverage: ${ENV_DIR}/bin/coverage cluster_mgr_setup.py
 	rm -rf htmlcov/
 	${ENV_DIR}/bin/coverage erase
-	${ENV_DIR}/bin/coverage run --source src --omit "**/test/**"  setup.py test
+	${ENV_DIR}/bin/coverage run --source src --omit "**/test/**"  cluster_mgr_setup.py test
 	${ENV_DIR}/bin/coverage report -m
 	${ENV_DIR}/bin/coverage html
 
 .PHONY: env
-env: setup.py $(ENV_DIR)/bin/python build-eggs install-eggs
+env: cluster_mgr_setup.py config_mgr_setup.py shared_setup.py $(ENV_DIR)/bin/python build-eggs install-eggs
 
 $(ENV_DIR)/bin/python:
 	# Set up the virtual environment
@@ -50,19 +50,34 @@ $(ENV_DIR)/bin/python:
 
 include build-infra/cw-deb.mk
 
-.PHONY: build-eggs
-build-eggs: setup.py common/setup.py src
+.PHONY: config-mgr-build-eggs
+config-mgr-build-eggs: config_mgr_setup.py shared_setup.py common/setup.py src
 	# Generate .egg files
-	${ENV_DIR}/bin/python setup.py bdist_egg -d eggs
-	cd common && ${ENV_DIR}/bin/python setup.py bdist_egg -d ../eggs
+	${ENV_DIR}/bin/python config_mgr_setup.py bdist_egg -d config_mgr_eggs
+	${ENV_DIR}/bin/python shared_setup.py bdist_egg -d config_mgr_eggs
+	cd common && ${ENV_DIR}/bin/python setup.py bdist_egg -d ../config_mgr_eggs
 
 	# Download the egg files they depend upon
-	${ENV_DIR}/bin/easy_install -zmaxd eggs/ eggs/*.egg
+	${ENV_DIR}/bin/easy_install -zmaxd config_mgr_eggs/ config_mgr_eggs/*.egg
+
+.PHONY: build-eggs
+build-eggs: cluster-mgr-build-eggs config-mgr-build-eggs
+
+.PHONY: cluster-mgr-build-eggs
+cluster-mgr-build-eggs: cluster_mgr_setup.py shared_setup.py common/setup.py src
+	# Generate .egg files
+	${ENV_DIR}/bin/python cluster_mgr_setup.py bdist_egg -d cluster_mgr_eggs
+	${ENV_DIR}/bin/python shared_setup.py bdist_egg -d cluster_mgr_eggs
+	cd common && ${ENV_DIR}/bin/python setup.py bdist_egg -d ../cluster_mgr_eggs
+
+	# Download the egg files they depend upon
+	${ENV_DIR}/bin/easy_install -zmaxd cluster_mgr_eggs/ cluster_mgr_eggs/*.egg
 
 .PHONY: install-eggs
-install-eggs: eggs/
+install-eggs: cluster_mgr_eggs config_mgr_eggs
 	# Install the downloaded egg files (this should match the Debian postinst)
-	${ENV_DIR}/bin/easy_install --allow-hosts=None -f eggs/ eggs/*.egg
+	${ENV_DIR}/bin/easy_install --allow-hosts=None -f cluster_mgr_eggs/ cluster_mgr_eggs/*.egg
+	${ENV_DIR}/bin/easy_install --allow-hosts=None -f config_mgr_eggs/ config_mgr_eggs/*.egg
 
 .PHONY: deb
 deb: env build-eggs deb-only
