+++ debian/clearwater-config-manager.install
@@ -1,2 +1,2 @@
-eggs usr/share/clearwater/clearwater-config-manager/
+config_mgr_eggs/* usr/share/clearwater/clearwater-config-manager/eggs/
 clearwater-config-manager.root/* /
