+++ shared_setup.py
@@ -1,3 +1,5 @@
+# @file setup.py
+#
 # Project Clearwater - IMS in the Cloud
 # Copyright (C) 2015 Metaswitch Networks Ltd
 #
@@ -30,30 +32,20 @@
 # under which the OpenSSL Project distributes the OpenSSL toolkit software,
 # as those licenses appear in the file LICENSE-OPENSSL.
 
-
-import subprocess
 import logging
+import sys
+import multiprocessing
 
-_log = logging.getLogger("config_manager.plugin_utils")
+from setuptools import setup, find_packages
 
-
-def run_command(command):
-    """Runs the given shell command, logging the output and return code.
-
-    Note that this runs the provided command in a new shell, which will
-    apply shell replacements.  Ensure the input string is sanitized before
-    passing to this function.
-    """
-    try:
-        output = subprocess.check_output(command,
-                                         shell=True,
-                                         stderr=subprocess.STDOUT)
-        _log.info("Command {} succeeded and printed output {!r}".
-                  format(command, output))
-        return 0
-    except subprocess.CalledProcessError as e:
-        _log.error("Command {} failed with return code {}"
-                   " and printed output {!r}".format(command,
-                                                     e.returncode,
-                                                     e.output))
-        return e.returncode
+setup(
+    name='clearwater-etcd-shared',
+    version='1.0',
+    packages=['metaswitch.clearwater.etcd_shared'],
+    package_dir={'':'src'},
+    package_data={
+        '': ['*.eml'],
+        },
+    install_requires=["docopt", "python-etcd", "pyzmq", "pyyaml"],
+    tests_require=["Mock"],
+    )
