+++ src/metaswitch/clearwater/config_manager/__init__.py
@@ -1,0 +1,1 @@
+from metaswitch.clearwater.etcd_shared import plugin_utils
