+++ src/metaswitch/clearwater/config_manager/main.py
@@ -51,7 +51,7 @@ Options:
 from docopt import docopt
 
 from metaswitch.common import logging_config, utils
-from metaswitch.clearwater.config_manager.plugin_loader \
+from metaswitch.clearwater.etcd_shared.plugin_loader \
     import load_plugins_in_dir
 from metaswitch.clearwater.config_manager.etcd_synchronizer \
     import EtcdSynchronizer
